+++ app/client-app/src/app/pages/Wordlist/index.jsx
@@ -1,102 +1,20 @@
-import { useEffect, useState } from "react";
-import {
-  Stack,
-  Flex,
-  RangeSlider,
-  RangeSliderTrack,
-  RangeSliderFilledTrack,
-  RangeSliderThumb,
-  HStack,
-  RangeSliderMark,
-  useRadioGroup,
-  Text,
-  VStack,
-  Spacer,
-} from "@chakra-ui/react";
-import { excludeBuiltinWords, includeBuiltinWords } from "api/wordlist";
+import { useState } from "react";
+import { Stack, Flex, useRadioGroup, VStack, Spacer } from "@chakra-ui/react";
 import LightHeading from "components/core/LightHeading";
-import adjectives from "services/adjectives";
-import animals from "services/animals";
-import { fetchWordList } from "api/wordlist";
 import { Page } from "components/core";
-import { toDictionary } from "services/data-structures";
-import { getFilteredWordList } from "./components/helpers";
 import { useWordlistSortingAndFiltering } from "./components/useWordlistSortingAndFiltering";
 import { WordCard } from "./components/WordCard";
 import WordlistSortingAndFilteringPanel from "./WordlistSortingAndFiltering";
 import { FixedSizeList } from "react-window";
 import AutoSizer from "react-virtualized-auto-sizer";
-import { RadioCard } from "./components/RadioCard";
+import {
+  ExclusionFilter,
+  TypeFilter,
+  WordLengthFilter,
+} from "./components/WordCardFilters";
+import { useWordData } from "./components/useWordData";
 
-const Wordlist = () => {
-  const [wordlist, setWordlist] = useState(null);
-  const [cards, setCards] = useState([]);
-  const { sorting, onSort, outputList, filter, setFilter } =
-    useWordlistSortingAndFiltering(cards);
-  const [sliderValues, setSliderValues] = useState([1, 15]);
-
-  const handleSliderChange = (values) => {
-    setSliderValues(values);
-  };
-
-  const typeOptions = ["Adjective", "Noun", "All"];
-  const exclusionState = ["Excluded", "Included", "All"];
-
-  const { getRootProps, getRadioProps } = useRadioGroup({
-    name: "framework",
-    defaultValue: "react",
-    onChange: console.log,
-  });
-
-  const group = getRootProps();
-
-  useEffect(() => {
-    const getWordList = async () => {
-      const data = await fetchWordList();
-      setWordlist(data);
-    };
-
-    getWordList();
-  }, []);
-
-  useEffect(() => {
-    if (wordlist) {
-      const excludedBuiltinsDict = toDictionary(
-        wordlist.excludedBuiltins,
-        "word"
-      );
-      const adjectiveCards = getFilteredWordList(
-        adjectives,
-        excludedBuiltinsDict,
-        "adjective"
-      );
-      const nounCards = getFilteredWordList(
-        animals,
-        excludedBuiltinsDict,
-        "noun"
-      );
-      setCards([...adjectiveCards, ...nounCards]);
-    }
-  }, [wordlist]);
-
-  const toggleExclude = async (word, type, isExcludedBuiltin) => {
-    const id = wordlist.id;
-
-    if (isExcludedBuiltin) {
-      await includeBuiltinWords(id, type, word);
-    } else {
-      await excludeBuiltinWords(id, type, word);
-    }
-
-    setCards((prevCards) =>
-      prevCards.map((card) =>
-        card.word === word
-          ? { ...card, isExcludedBuiltin: !isExcludedBuiltin }
-          : card
-      )
-    );
-  };
-
+const WordlistDisplay = ({ outputList, height, width, toggleExclude }) => {
   const RenderWordCard = ({ index, style }) => {
     const card = outputList[index];
     return card ? (
@@ -112,6 +30,37 @@ const Wordlist = () => {
       </div>
     ) : null;
   };
+
+  return (
+    <FixedSizeList
+      height={height}
+      width={width}
+      itemCount={outputList.length}
+      itemSize={80}
+    >
+      {RenderWordCard}
+    </FixedSizeList>
+  );
+};
+
+const Wordlist = () => {
+  const { cards, toggleExclude } = useWordData();
+  const { sorting, onSort, outputList, filter, setFilter } =
+    useWordlistSortingAndFiltering(cards);
+  const [sliderValues, setSliderValues] = useState([1, 15]);
+
+  const handleSliderChange = (values) => {
+    setSliderValues(values);
+  };
+
+  const { getRootProps, getRadioProps } = useRadioGroup({
+    name: "framework",
+    defaultValue: "react",
+    onChange: console.log,
+  });
+
+  const group = getRootProps();
+
   return (
     <Page layout="wordlist">
       <Flex direction="column" height={`calc(100vh - 54px)`} width="100%">
@@ -121,64 +70,12 @@ const Wordlist = () => {
         <Stack mt={2} spacing={4} h="100vh" p={2}>
           <Flex p={2} boxShadow="base">
             <VStack alignItems="start">
-              <HStack {...group}>
-                <Text>Type:</Text>
-                {typeOptions.map((value) => {
-                  const radio = getRadioProps({ value });
-                  return (
-                    <RadioCard key={value} {...radio}>
-                      {value}
-                    </RadioCard>
-                  );
-                })}
-              </HStack>
-              <HStack {...group}>
-                <Text>State:</Text>
-                {exclusionState.map((value) => {
-                  const radio = getRadioProps({ value });
-                  return (
-                    <RadioCard key={value} {...radio}>
-                      {value}
-                    </RadioCard>
-                  );
-                })}
-              </HStack>
-              <HStack pt={4}>
-                <Text>Word Length:</Text>
-                <Flex width="300px">
-                  <RangeSlider
-                    id="word-length"
-                    defaultValue={[1, 15]}
-                    min={1}
-                    max={15}
-                    step={1}
-                    onChange={handleSliderChange}
-                    value={sliderValues}
-                  >
-                    <RangeSliderMark
-                      value={sliderValues[0]}
-                      textAlign="center"
-                      mt="-30px"
-                      ml="-6px"
-                    >
-                      {sliderValues[0]}
-                    </RangeSliderMark>
-                    <RangeSliderMark
-                      value={sliderValues[1]}
-                      textAlign="center"
-                      mt="-30px"
-                      ml="-6px"
-                    >
-                      {sliderValues[1]}
-                    </RangeSliderMark>
-                    <RangeSliderTrack bg="blue.100">
-                      <RangeSliderFilledTrack bg="blue.500" />
-                    </RangeSliderTrack>
-                    <RangeSliderThumb boxSize={4} index={0} />
-                    <RangeSliderThumb boxSize={4} index={1} />
-                  </RangeSlider>
-                </Flex>
-              </HStack>
+              <TypeFilter group={group} getRadioProps={getRadioProps} />
+              <ExclusionFilter group={group} getRadioProps={getRadioProps} />
+              <WordLengthFilter
+                sliderValues={sliderValues}
+                handleSliderChange={handleSliderChange}
+              />
             </VStack>
             <Spacer />
             <WordlistSortingAndFilteringPanel
@@ -189,18 +86,15 @@ const Wordlist = () => {
               setFilter={setFilter}
             />
           </Flex>
-
           <Flex flex="1">
             <AutoSizer>
               {({ height, width }) => (
-                <FixedSizeList
+                <WordlistDisplay
+                  outputList={outputList}
                   height={height}
                   width={width}
-                  itemCount={outputList.length}
-                  itemSize={80}
-                >
-                  {RenderWordCard}
-                </FixedSizeList>
+                  toggleExclude={toggleExclude}
+                />
               )}
             </AutoSizer>
           </Flex>
@@ -209,4 +103,5 @@ const Wordlist = () => {
     </Page>
   );
 };
+
 export default Wordlist;
